@@ -21,5 +21,6 @@ bind = app_config.host + ':' + str(app_config.port)
 # We use 'os.sched_getaffinity(pid)' not 'os.cpu_count()' because it returns only allowable CPUs.
 # os.sched_getaffinity(pid): Return the set of CPUs the process with PID pid is restricted to.
 # os.cpu_count(): Return the number of CPUs in the system.
-workers = len(os.sched_getaffinity(0)) * 2
+
+workers = (len(os.sched_getaffinity(0)) * 2) + 1
 threads = 4
