@@ -1,32 +1,58 @@
 import logging
 import os
-from typing import Optional
+from typing import Any, Tuple
 
-from apiflask import APIFlask
-from flask import g
-from werkzeug.exceptions import Unauthorized
+from apiflask import APIFlask, exceptions
+from flask_cors import CORS
+from pydantic import Field
 
 import src.adapters.db as db
 import src.adapters.db.flask_db as flask_db
+import src.api.feature_flags.feature_flag_config as feature_flag_config
 import src.logging
 import src.logging.flask_logger as flask_logger
 from src.api.healthcheck import healthcheck_blueprint
+from src.api.opportunities_v0 import opportunity_blueprint as opportunities_v0_blueprint
+from src.api.opportunities_v0_1 import opportunity_blueprint as opportunities_v0_1_blueprint
+from src.api.response import restructure_error_response
 from src.api.schemas import response_schema
-from src.api.users import user_blueprint
-from src.auth.api_key_auth import User, get_app_security_scheme
+from src.auth.api_key_auth import get_app_security_scheme
+from src.data_migration.data_migration_blueprint import data_migration_blueprint
+from src.util.env_config import PydanticBaseEnvConfig
 
 logger = logging.getLogger(__name__)
 
+TITLE = "Simpler Grants API"
+API_OVERALL_VERSION = "v0"
+API_DESCRIPTION = """
+Back end API for simpler.grants.gov.
+
+This API is an ALPHA VERSION! Its current form is primarily for testing and feedback. Features are still under heavy development, and subject to change. Not for production use.
+
+See [Release Phases](https://github.com/github/roadmap?tab=readme-ov-file#release-phases) for further details.
+"""
+
+
+class EndpointConfig(PydanticBaseEnvConfig):
+    """
+    Class which holds environments variables for enabling/disabling API
+    endpoint blueprints. Useful while we develop new endpoints that we want
+    to keep hidden in production (or other) environments temporarily while
+    we build out the features.
+    """
+
+    enable_v_0_1_endpoints: bool = Field(False, alias="ENABLE_V_0_1_ENDPOINTS")
+
 
 def create_app() -> APIFlask:
-    app = APIFlask(__name__)
+    app = APIFlask(__name__, title=TITLE, version=API_OVERALL_VERSION)
 
-    src.logging.init(__package__)
-    flask_logger.init_app(logging.root, app)
+    setup_logging(app)
+    register_db_client(app)
 
-    db_client = db.PostgresDBClient()
-    flask_db.register_db_client(db_client, app)
+    feature_flag_config.initialize()
 
+    CORS(app)
     configure_app(app)
     register_blueprints(app)
     register_index(app)
@@ -34,12 +60,14 @@ def create_app() -> APIFlask:
     return app
 
 
-def current_user(is_user_expected: bool = True) -> Optional[User]:
-    current = g.get("current_user")
-    if is_user_expected and current is None:
-        logger.error("No current user found for request")
-        raise Unauthorized
-    return current
+def setup_logging(app: APIFlask) -> None:
+    src.logging.init(__package__)
+    flask_logger.init_app(logging.root, app)
+
+
+def register_db_client(app: APIFlask) -> None:
+    db_client = db.PostgresDBClient()
+    flask_db.register_db_client(db_client, app)
 
 
 def configure_app(app: APIFlask) -> None:
@@ -47,18 +75,27 @@ def configure_app(app: APIFlask) -> None:
     # which adds additional details to the object.
     # https://apiflask.com/schema/#base-response-schema-customization
     app.config["BASE_RESPONSE_SCHEMA"] = response_schema.ResponseSchema
+    app.config["HTTP_ERROR_SCHEMA"] = response_schema.ErrorResponseSchema
+    app.config["VALIDATION_ERROR_SCHEMA"] = response_schema.ErrorResponseSchema
+    app.config["SWAGGER_UI_CSS"] = "/static/swagger-ui.min.css"
+    app.config["SWAGGER_UI_BUNDLE_JS"] = "/static/swagger-ui-bundle.js"
+    app.config["SWAGGER_UI_STANDALONE_PRESET_JS"] = "/static/swagger-ui-standalone-preset.js"
+    # Removing because the server dropdown has accessibility issues.
+    app.config["SERVERS"] = "."
+    app.config["DOCS_FAVICON"] = "https://simpler.grants.gov/img/favicon.ico"
 
     # Set a few values for the Swagger endpoint
-    app.config["OPENAPI_VERSION"] = "3.0.3"
+    app.config["OPENAPI_VERSION"] = "3.1.0"
+
+    app.json.compact = False  # type: ignore
 
     # Set various general OpenAPI config values
     app.info = {
-        "title": "Grants Equity API",
-        "description": "Back end API for grants.gov",
+        "description": API_DESCRIPTION,
         "contact": {
-            "name": "Nava PBC Engineering",
-            "url": "https://www.navapbc.com",
-            "email": "engineering@navapbc.com",
+            "name": "Simpler Grants.gov",
+            "url": "https://simpler.grants.gov/",
+            "email": "simpler@grants.gov",
         },
     }
 
@@ -67,10 +104,23 @@ def configure_app(app: APIFlask) -> None:
     # See: https://apiflask.com/authentication/#use-external-authentication-library
     app.security_schemes = get_app_security_scheme()
 
+    @app.error_processor
+    def error_processor(error: exceptions.HTTPError) -> Tuple[dict, int, Any]:
+        return restructure_error_response(error)
+
 
 def register_blueprints(app: APIFlask) -> None:
+    endpoint_config = EndpointConfig()
+
     app.register_blueprint(healthcheck_blueprint)
-    app.register_blueprint(user_blueprint)
+    app.register_blueprint(opportunities_v0_blueprint)
+    if endpoint_config.enable_v_0_1_endpoints:
+        logger.info("Enabling v0.1 endpoints")
+        app.register_blueprint(opportunities_v0_1_blueprint)
+    else:
+        logger.info("v0.1 endpoints are not enabled")
+
+    app.register_blueprint(data_migration_blueprint)
 
 
 def get_project_root_dir() -> str:
