@@ -1,17 +1,19 @@
 import logging
 import os
-import uuid
 from dataclasses import dataclass
 from typing import Any
 
 import flask
-from apiflask import HTTPTokenAuth, abort
+from apiflask import HTTPTokenAuth
+
+from src.api.route_utils import raise_flask_error
+from src.logging.flask_logger import add_extra_data_to_current_request_logs
 
 logger = logging.getLogger(__name__)
 
 # Initialize the authorization context
 # this needs to be attached to your
-# routes as `your_blueprint..auth_required(api_key_auth)`
+# routes as `your_blueprint.auth_required(api_key_auth)`
 # in order to enable authorization
 api_key_auth = HTTPTokenAuth("ApiKey", header="X-Auth")
 
@@ -22,31 +24,16 @@ def get_app_security_scheme() -> dict[str, Any]:
 
 @dataclass
 class User:
-    # WARNING: This is a very rudimentary
-    # user for demo purposes and is not
-    # a production ready approach. It exists
-    # purely to define a rough structure / example
-    id: uuid.UUID
-    sub_id: str
+    # A very basic "user" for the purposes of logging
+    # which API key was used to call us.
     username: str
 
-    def as_dict(self) -> dict[str, Any]:
-        # Connexion expects a dictionary it can
-        # use .get() on, so convert this to that format
-        return {"uid": self.id, "sub": self.sub_id}
 
-    def get_user_log_attributes(self) -> dict:
-        # Note this gets called during authentication
-        # to attach the information to the flask global object
-        # which will in turn be attached to the log record
-        return {"current_user.id": str(self.id)}
-
-
-API_AUTH_USER = User(uuid.uuid4(), "sub_id_1234", "API auth user")
+API_AUTH_USERS: dict[str, User] | None = None
 
 
 @api_key_auth.verify_token
-def verify_token(token: str) -> dict:
+def verify_token(token: str) -> User:
     logger.info("Authenticating provided token")
 
     user = process_token(token)
@@ -58,11 +45,46 @@ def verify_token(token: str) -> dict:
     # in your API, you don't need to check each
     # one in order to figure out which was actually used
     flask.g.current_user = user
-    flask.g.current_user_log_attributes = user.get_user_log_attributes()
+
+    # Add the "username" to all logs for the rest of the request lifecycle
+    add_extra_data_to_current_request_logs({"auth.username": user.username})
 
     logger.info("Authentication successful")
 
-    return user.as_dict()
+    return user
+
+
+def _get_auth_users() -> dict[str, User] | None:
+    # To avoid every request re-parsing the auth token
+    # string, load it once and store in a global variable
+    # This doesn't really account for threading, but worst case
+    # multiple threads write the same value
+    global API_AUTH_USERS
+
+    if API_AUTH_USERS is not None:
+        return API_AUTH_USERS
+
+    raw_auth_tokens = os.getenv("API_AUTH_TOKEN", None)
+
+    if raw_auth_tokens is None:
+        return None
+
+    API_AUTH_USERS = {}
+
+    # Auth tokens will look like some_value,another_value,a_third_value
+    # and get usernames like:
+    # some_value    -> auth_token_0
+    # another_value -> auth_token_1
+    # a_third_value -> auth_token_2
+    #
+    # This username is just used for logging, and is deliberately very
+    # simple until we build out authentication more. This just gives us the most
+    # barebones ability to do distinguish our users.
+    auth_tokens = raw_auth_tokens.split(",")
+    for i, token in enumerate(auth_tokens):
+        API_AUTH_USERS[token] = User(username=f"auth_token_{i}")
+
+    return API_AUTH_USERS
 
 
 def process_token(token: str) -> User:
@@ -71,18 +93,22 @@ def process_token(token: str) -> User:
     # here should be pulled from a DB or auth service, but
     # as there are several types of authentication, we are
     # keeping this pretty basic for the out-of-the-box approach
-    expected_auth_token = os.getenv("API_AUTH_TOKEN", None)
+    api_auth_users = _get_auth_users()
 
-    if not expected_auth_token:
-        logger.info(
+    if not api_auth_users:
+        logger.error(
             "Authentication is not setup, please add an API_AUTH_TOKEN environment variable."
         )
-        abort(401, "Authentication is not setup properly and the user cannot be authenticated")
+        raise_flask_error(
+            401, "Authentication is not setup properly and the user cannot be authenticated"
+        )
 
-    if token != expected_auth_token:
+    user = api_auth_users.get(token, None)
+
+    if user is None:
         logger.info("Authentication failed for provided auth token.")
-        abort(
+        raise_flask_error(
             401, "The server could not verify that you are authorized to access the URL requested"
         )
 
-    return API_AUTH_USER
+    return user
