@@ -1,6 +1,7 @@
 import logging
 
-from . import base, user_models
+from . import base, lookup_models, opportunity_models
+from .transfer import topportunity_models
 
 logger = logging.getLogger(__name__)
 
@@ -8,4 +9,4 @@ logger = logging.getLogger(__name__)
 # This is used by tests to create the test database.
 metadata = base.metadata
 
-__all__ = ["metadata", "user_models"]
+__all__ = ["metadata", "opportunity_models", "lookup_models", "topportunity_models"]
