@@ -1,5 +1,8 @@
+import { ExternalRoutes } from "src/constants/routes";
+
 import { useTranslation } from "next-i18next";
-import { GridContainer } from "@trussworks/react-uswds";
+import Link from "next/link";
+import { GridContainer, Icon } from "@trussworks/react-uswds";
 
 const Hero = () => {
   const { t } = useTranslation("common", {
@@ -7,18 +10,26 @@ const Hero = () => {
   });
 
   return (
-    <div
-      data-testid="hero"
-      className="bg-primary desktop:padding-y-15 tablet-lg:padding-y-10 tablet:padding-y-3 padding-y-1"
-    >
-      <GridContainer>
-        <h1 className="text-white desktop:font-sans-3xl tablet:font-sans-2xl font-sans-xl">
-          {t("title")}
+    <div data-testid="hero" className="usa-dark-background bg-primary">
+      <GridContainer className="padding-y-1 tablet:padding-y-3 tablet-lg:padding-y-10 desktop-lg:padding-y-15 position-relative">
+        <h1 className="text-ls-neg-2 tablet:font-sans-2xl desktop-lg:font-sans-3xl desktop-lg:margin-top-2">
+          <span>{t("title")}</span>
         </h1>
-        <p className="usa-intro desktop:font-sans-xl tablet:font-sans-lg font-sans-md">
-          <span className="text-yellow text-bold">{t("beta")}</span>&nbsp;
-          <span className="text-white">{t("content")}</span>
+        <p className="usa-intro line-height-sans-3 font-sans-md tablet:font-sans-lg desktop-lg:font-sans-xl desktop-lg:margin-bottom-0">
+          {t("content")}
         </p>
+        <Link
+          className="usa-button usa-button--outline usa-button--inverse font-sans-2xs tablet:font-sans-md margin-bottom-3 desktop:position-absolute top-3 right-0 desktop:margin-y-3 desktop:margin-x-4"
+          href={ExternalRoutes.GITHUB_REPO}
+          target="_blank"
+        >
+          <Icon.Github
+            className="usa-icon usa-icon--size-3 margin-right-1 text-middle tablet:margin-top-neg-2px"
+            size={3}
+            aria-label="Github"
+          />
+          {t("github_link")}
+        </Link>
       </GridContainer>
     </div>
   );
